Rename CustomService to DocBlockService
